Topology Manager Update - socketmask package builds SocketMask
