+++ pkg/kubelet/cm/topologymanager/topologymanager.go
@@ -12,10 +12,7 @@
  
 package topologymanager
 import (
- 	"strconv"
- 	"strings"
-	"bytes"
-	"math"
+	"k8s.io/kubernetes/pkg/kubelet/cm/topologymanager/socketmask"
  	"github.com/golang/glog"	
  	"k8s.io/api/core/v1"
 	"k8s.io/kubernetes/pkg/kubelet/lifecycle"
@@ -28,12 +25,8 @@ type Manager interface {
  	Store
  }
 
-type SocketMask struct {
-        Mask [][]int64
-}
-
 type TopologyHints struct {
-	SocketAffinity SocketMask
+	SocketAffinity socketmask.SocketMask
 	Affinity bool
 }
  
@@ -92,12 +85,9 @@ func (m *manager) GetAffinity(podUID string, containerName string) TopologyHints
 }
 
 func (m *manager) calculateTopologyAffinity(pod v1.Pod, container v1.Container) TopologyHints {
-	socketMask := SocketMask{
-		Mask:		nil,
-	}
-	
+	sm := socketmask.NewSocketMask(nil)
 	podTopologyHints := TopologyHints {
-		SocketAffinity:	socketMask,
+		SocketAffinity:	sm,
 		Affinity: 	true,
 	}
 		
@@ -110,16 +100,16 @@ func (m *manager) calculateTopologyAffinity(pod v1.Pod, container v1.Container) 
 			topologyHints := hp.GetTopologyHints(string(resource), int(amount.Value()))
 			if topologyHints.Affinity && topologyHints.SocketAffinity.Mask != nil {
 				if count == 0 {
-					maskHolder = buildMaskHolder(topologyHints.SocketAffinity.Mask)
+					maskHolder = sm.BuildMaskHolder(topologyHints.SocketAffinity.Mask)
 					count++
 				}
 				glog.Infof("[topologymanager] MaskHolder : %v", maskHolder)
 				//Arrange int array into array of strings 
 				glog.Infof("[topologymanager] %v is passed into arrange function",topologyHints.SocketAffinity.Mask)   
-				arrangedMask := arrangeMask(topologyHints.SocketAffinity.Mask)						
-				newMask := getTopologyAffinity(arrangedMask, maskHolder)
+				arrangedMask := sm.ArrangeMask(topologyHints.SocketAffinity.Mask)						
+				newMask := sm.GetTopologyAffinity(arrangedMask, maskHolder)
 				glog.Infof("[topologymanager] New Mask after getTopologyAffinity (new mask) : %v ",newMask)
-				finalMaskValue = parseMask(newMask)
+				finalMaskValue = sm.ParseMask(newMask)
 				glog.Infof("[topologymanager] Mask []Int64 (finalMaskValue): %v", finalMaskValue)
 				maskHolder = newMask
 				glog.Infof("[topologymanager] New MaskHolder: %v", maskHolder) 
@@ -136,119 +126,6 @@ func (m *manager) calculateTopologyAffinity(pod v1.Pod, container v1.Container) 
         podTopologyHints.SocketAffinity.Mask = topologyMaskFull
         return podTopologyHints      
 
-}
-func buildMaskHolder(mask [][]int64) []string {
-	var maskHolder []string
-	outerLen := len(mask)
-        var innerLen int = 0 
-      	for i := 0; i < outerLen; i++ {
-        	if innerLen < len(mask[i]) {
-          		innerLen = len(mask[i])
-       		}
-     	}
-       	var buffer bytes.Buffer
-   	var i, j int = 0, 0
-       	for i = 0; i < outerLen; i++ {
-    		for j = 0; j < innerLen; j++ {
-            		buffer.WriteString("1")
-     		}
-         	maskHolder = append(maskHolder, buffer.String())
-               	buffer.Reset()
-     	}
-	return maskHolder
-}
-
-func getTopologyAffinity(arrangedMask, maskHolder []string) []string {
-	var topologyTemp []string
-        for i:= 0; i < (len(maskHolder)); i++ {
-        	for j:= 0; j < (len(arrangedMask)); j++ {
-               		tempStr := andOperation(maskHolder[i],arrangedMask[j])
-                      	if strings.Contains(tempStr, "1") {
-                               	topologyTemp = append(topologyTemp, tempStr )
-                      	}
-               	}
-     	}
-        duplicates := map[string]bool{}
-        for v:= range topologyTemp {
-        	duplicates[topologyTemp[v]] = true
-        }
-       	// Place all keys from the map into a slice.
-       	topologyResult := []string{}
-      	for key, _ := range duplicates {
-       		topologyResult = append(topologyResult, key)
-     	}
-	
-	return topologyResult
-}
-
-func parseMask(mask []string) []int64 {
-	var maskStr string
-	min := strings.Count(mask[0], "1")
-	var num, index int
-	
-       	for i := 0; i < len(mask); i++ {
-		num = strings.Count(mask[i], "1")
-		if num < min {
-			min = num
-			index = i
-		} 
-		maskStr = mask[index]
-        }
-	var maskInt []int64
-	for _, char := range maskStr {
-		convertedStr, err := strconv.Atoi(string(char))
-		if err != nil {
-			glog.Errorf("Could not convert string to int. Err: %v", err)
-			return maskInt
-		}
-		maskInt = append(maskInt, int64(convertedStr))  
-	}
-	glog.Infof("Mask Int in Parse Mask: %v", maskInt)
-	return maskInt         
-}
-
-func arrangeMask(mask [][]int64) []string {
-	var socketStr []string
-	var bufferNew bytes.Buffer
-	outerLen := len(mask)
-	innerLen := len(mask[0])
-	for i := 0; i < outerLen; i++ {
-		for j := 0; j < innerLen; j++ {
-			if mask[i][j] == 1 {
-				bufferNew.WriteString("1")
-			} else if mask[i][j] == 0 {
-				bufferNew.WriteString("0")
-			}
-		}
-		socketStr = append(socketStr, bufferNew.String())
-		bufferNew.Reset()
-	}
-	return socketStr
-}
-
-func andOperation(val1, val2 string) (string) {
-	l1, l2 := len(val1), len(val2)
-	//compare lengths of strings - pad shortest with trailing zeros
-	if l1 != l2 {
-		// Get the bit difference
-		var num int
-		diff := math.Abs(float64(l1) - float64(l2))
-		num = int(diff)
-		if l1 < l2 {
-			val1 = val1 + strings.Repeat("0", num)
-		} else {
-			val2 = val2 + strings.Repeat("0", num)
-		}
-	}
-	length := len(val1)
-	byteArr := make([]byte, length)
-	for i := 0; i < length ; i++ {
-		byteArr[i] = (val1[i] & val2[i])
-    	}
-	var finalStr string
-	finalStr = string(byteArr[:])	
-	
-	return finalStr
 }
 
 func (m *manager) AddHintProvider(h HintProvider) {
